move background color to theme, add givingful img assets and use em, move header and footer into components folder, replace bell icon with mui icon
