+++ client/src/Theme.tsx
@@ -47,6 +47,9 @@ const theme = createTheme({
     },
   },
   palette: {
+    background: {
+      default: '#FFFFFF',
+    },
     text: {
       primary: '#000000',
       secondary: '#999999',
